# docs/swagger.yaml
definitions:
  helper.CreateSosmedResponse:
    properties:
      id:
        type: integer
      message:
        type: string
      social_media_url:
        type: string
      sosmed:
        type: string
      user_id:
        type: integer
    type: object
  helper.ErrorResponse:
    properties:
      err:
        type: string
      message:
        type: string
    type: object
  helper.SuccessResponse:
    properties:
      data: {}
      message:
        type: string
      status:
        type: integer
    type: object
  helper.UpdateSosmedResponse:
    properties:
      id:
        type: integer
      message:
        type: string
      social_media_url:
        type: string
      sosmed:
        type: string
      user_id:
        type: integer
    type: object
  helper.UserLoginResponse:
    properties:
      message:
        type: string
      token:
        type: string
    type: object
  helper.UserRegisterResponse:
    properties:
      email:
        type: string
      full_name:
        type: string
      id:
        type: integer
      message:
        type: string
    type: object
...
  model.LoginReq:
    properties:
      email:
        example: budi@sep.com
        type: string
      password:
        example: budi69
        type: string
    type: object
...
  model.SosmedReq:
    properties:
      name:
        example: Facebook
        type: string
      social_media_url:
        example: facebook.com/budi_anduk
        type: string
    type: object
...
  model.UpdateSosmedReq:
    properties:
      name:
        example: Facebook Lite
        type: string
      social_media_url:
        example: facebook.com/budi_anduk_bunduk
        type: string
    type: object
  model.UserReqEq:
    properties:
      age:
        example: 22
        type: integer
      email:
        example: budi@sep.com
        type: string
      password:
        example: budi69
        type: string
      user_name:
        example: budianduk
        type: string
    type: object
info:
  contact: {}
paths:
...
  /sosmed:
    get:
      consumes:
      - application/json
      description: Get all Sosmed
      parameters:
      - description: Authorization header in the format of 'Bearer {token}'
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helper.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get all Sosmed
      tags:
      - Social Media
    post:
      consumes:
      - application/json
      description: Create a new social media account
      parameters:
      - description: Authorization header in the format of 'Bearer {token}'
        in: header
        name: Authorization
        required: true
        type: string
      - description: Sosmed data
        in: body
        name: sosmed
        required: true
        schema:
          $ref: '#/definitions/model.SosmedReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/helper.CreateSosmedResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create a new social media account
      tags:
      - Social Media
  /sosmed/{sosmedId}:
    delete:
      consumes:
      - application/json
      description: Delete a social media by ID
      parameters:
      - description: Authorization header in the format of 'Bearer {token}'
        in: header
        name: Authorization
        required: true
        type: string
      - description: Sosmed ID
        in: path
        name: sosmedId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete a social media by ID
      tags:
      - Social Media
    get:
      consumes:
      - application/json
      description: Get Sosmed by ID
      parameters:
      - description: Authorization header in the format of 'Bearer {token}'
        in: header
        name: Authorization
        required: true
        type: string
      - description: Sosmed ID
        in: path
        name: sosmedId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helper.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get Sosmed by ID
      tags:
      - Social Media
    put:
      consumes:
      - application/json
      description: Update existing sosmed
      parameters:
      - description: Authorization header in the format of 'Bearer {token}'
        in: header
        name: Authorization
        required: true
        type: string
      - description: Sosmed ID
        in: path
        name: sosmedId
        required: true
        type: integer
      - description: Update sosmed object
        in: body
        name: sosmed
        required: true
        schema:
          $ref: '#/definitions/model.UpdateSosmedReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helper.UpdateSosmedResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update existing sosmed
      tags:
      - Social Media
  /users/login:
    post:
      consumes:
      - application/json
      description: User login API
      parameters:
      - description: User login credentials
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.LoginReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helper.UserLoginResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
      summary: User login
      tags:
      - Action For Users
  /users/register:
    post:
      consumes:
      - application/json
      description: Register a new user
      parameters:
      - description: User object that needs to be registered
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.UserReqEq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/helper.UserRegisterResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
      summary: Register a new user
      tags:
      - Action For Users
swagger: "2.0"